fix: resolve failing tests and improve translation key parity
- Add missing translation keys across all language files (es, fr, zh, hi)
  - Added setupSubtitle, customizationFeature, privacyFeature, varietyFeature
  - Added updatedGameSettings, updatedRoomSettings, updatedMultipleSettings
- Remove orphaned translation keys not present in English
  - Removed bugsTitle/Description, coachTitle/Description, couplesTitle/Description
- Fix MessageList component tests expecting tablist role
  - Updated test assertions to match FAB filter implementation
- Fix TypeScript TFunction import error in messageUtils
- Extract message utility functions for better code organization
- Improve CSS custom properties and spacing consistency
- Add accessibility improvements to focus states

All tests now pass (631 total) with proper translation key parity
