+++ src/components/MessageList/styles.css
@@ -44,7 +44,7 @@
 }
 
 .message {
-  padding: 12px 16px;
+  padding: var(--spacing-sm) var(--spacing-md);
   margin-bottom: 1px;
   background: transparent;
   border-bottom: 1px solid rgba(255, 255, 255, 0.1);
