+++ src/components/RoomBackground/index.tsx
@@ -7,13 +7,19 @@ interface RoomBackgroundProps {
   isVideo?: boolean | null;
 }
 
+// Enum for background types
+enum BackgroundType {
+  COLOR = '/images/color',
+  GRAY = '/images/gray',
+}
+
 export default function RoomBackground({ url = null, isVideo = null }: RoomBackgroundProps) {
   // Check if the URL is a direct video file (MP4, WebM, etc.)
   const isDirectVideo = url && /\.(mp4|webm|ogg|mov)(\?.*)?$/i.test(url);
 
   // Show default background when no custom background is set OR when background is "color" or "gray"
   const isNonImageBackground =
-    url && (url.includes('/images/color') || url.includes('/images/gray'));
+    url && (url.includes(BackgroundType.COLOR) || url.includes(BackgroundType.GRAY));
   const hasCustomBackground = url && !isNonImageBackground && (isVideo || (!isVideo && url));
 
   return (
