+++ src/locales/es/translation.json
@@ -22,6 +22,7 @@
   "tilesCount": "{{count}} casillas",
   "online": "En línea",
   "setup": "Configuración del juego",
+  "setupSubtitle": "Ingresa tu nombre y comienza la sesión",
   "buildGame": "Construir Juego",
   "access": "Acceder",
   "setupWizard": "Asist. Config.",
@@ -197,15 +198,13 @@
   "pickSomething": "Por favor selecciona al menos una acción de la pestaña Tablero antes de actualizar el juego.",
   "appendWithAction": "Las opciones de añadir requieren al menos una acción independiente. Por favor selecciona una acción que no sea de añadir en la pestaña Tablero.",
   "gameDesc": "<0>Blitzed Out</0><1></1><2>Un juego de mesa en línea para adultos, dirigido a personas fetichistas. Diseñado para jugar en solitario, en pareja o en fiestas.</2><3>Personaliza tu propio tablero o importa el de otra persona, encuentra tu propia sala y juega con otros en línea.</3>",
+  "whySpecialTitle": "¿Qué lo hace especial?",
+  "customizationFeature": "<0>🎲</0><1>Personalización ilimitada</1><2>Crea tu experiencia perfecta con tableros personalizados, acciones y niveles de intensidad adaptados a tus deseos.</2>",
+  "privacyFeature": "<0>🔒</0><1>Control total de privacidad</1><2>Salas privadas, opciones de juego local y sesiones anónimas. Tu privacidad y comodidad son lo primero.</2>",
+  "varietyFeature": "<0>🌟</0><1>Variedad infinita</1><2>Desde exploración suave hasta aventuras intensas: con contenido creado por la comunidad y actualizaciones regulares.</2>",
   "gettingStartedTitle": "Primeros pasos",
   "gettingStartedMainText": "Ingresa tu nombre arriba para comenzar tu sesión. Personaliza tu experiencia en el asistente de configuración. Tira los dados y deja que comience la aventura.",
   "gettingStartedTip": "¿Nuevo en esto? Comienza con niveles de intensidad más bajos y explora a tu propio ritmo. Siempre puedes ajustar la configuración en cualquier momento.",
-  "couplesTitle": "Parejas y Fiestas",
-  "couplesDescription": "<0>¿Jugando en la misma sala con tu pareja o amigos?</0><1>Ajustes > Modo de juego > Sala privada > Partida local > Actualizar</1><2>Nota: Si no deseas interactuar con las personas en la sala, elige Individual.</2>",
-  "coachTitle": "Entrenamiento automático",
-  "coachDescription": "<0>Ajustes > Modo de juego > Leer en alto mis tiradas > Actualizar. A continuación, haga clic en la flecha hacia arriba junto a ‘Tirar’ y seleccione el intervalo para las tiradas automáticas.</0><1>El juego hará las tiradas automáticamente y leerá en alto la acción que debe realizar.</1>",
-  "bugsTitle": "Errores, Comentarios y Sugerencias",
-  "bugsDescription": "Compártelo en nuestro <0>servidor de Discord</0>",
   "add": "Añadir",
   "delete": "Eliminar",
   "save": "Guardar",
@@ -518,5 +517,8 @@
   "Authentication": "Autenticación",
   "minutes": "minutos",
   "customTiles.actions": "Acciones",
-  "customUpdated": "Personalizado Actualizado"
+  "customUpdated": "Personalizado Actualizado",
+  "updatedGameSettings": "actualizó la configuración del juego",
+  "updatedRoomSettings": "actualizó la configuración de la sala",
+  "updatedMultipleSettings": "actualizó {{count}} configuraciones"
 }
