+++ src/locales/fr/translation.json
@@ -22,6 +22,7 @@
   "tilesCount": "{{count}} tuiles",
   "online": "En ligne",
   "setup": "Configuration du jeu",
+  "setupSubtitle": "Entrez votre nom et commencez la session",
   "buildGame": "Créer un Jeu",
   "access": "Accéder au jeu",
   "setupWizard": "Assistant",
@@ -197,15 +198,13 @@
   "pickSomething": "Veuillez sélectionner au moins une action dans l'onglet Plateau avant de mettre à jour le jeu.",
   "appendWithAction": "Les options d'ajout nécessitent au moins une action autonome. Veuillez sélectionner une action non-ajout dans l'onglet Plateau.",
   "gameDesc": "<0>Blitzed Out</0><1></1><2>Un jeu de société en ligne pour adultes coquins. Conçu pour jouer seul, en couple ou en groupe.</2><3>Personnalisez votre propre plateau de jeu ou importez celui de quelqu'un d'autre, trouvez une salle et jouez avec d'autres personnes en ligne.</3>",
+  "whySpecialTitle": "Qu'est-ce qui le rend spécial?",
+  "customizationFeature": "<0>🎲</0><1>Personnalisation illimitée</1><2>Créez votre expérience parfaite avec des plateaux personnalisés, des actions et des niveaux d'intensité adaptés à vos désirs.</2>",
+  "privacyFeature": "<0>🔒</0><1>Contrôle complet de la confidentialité</1><2>Salles privées, options de jeu local et sessions anonymes. Votre vie privée et votre confort passent en premier.</2>",
+  "varietyFeature": "<0>🌟</0><1>Variété infinie</1><2>De l'exploration douce aux aventures intenses — avec du contenu créé par la communauté et des mises à jour régulières.</2>",
   "gettingStartedTitle": "Comment commencer",
   "gettingStartedMainText": "Entrez votre nom ci-dessus pour commencer votre session. Personnalisez votre expérience dans l'assistant de configuration. Lancez les dés et laissez l'aventure commencer.",
   "gettingStartedTip": "Nouveau dans ce domaine ? Commencez avec des niveaux d'intensité plus légers et explorez à votre rythme. Vous pouvez toujours ajuster les paramètres à tout moment.",
-  "couplesTitle": "Couples et Groupes",
-  "couplesDescription": "<0>Vous jouez dans la même pièce avec un partenaire ou des amis en présentiel ?</0><1>Paramètres > Application > Salle privée > Jeu local.</1><2>Remarque: Si vous ne voulez pas interagir avec les personnes dans la salle, choisissez Solo.</2>",
-  "coachTitle": "Entraînement par robot",
-  "coachDescription": "<0>Paramètres > Application > Enonces mes jets de dés > Enregistrer / Accéder au jeu. Ensuite, cliquez sur la flèche vers le haut à côté du dé et choisissez l'intervalle de lancer automatique.</0><1>Le jeu lancera automatiquement le dé et vous indiquera l'action à effectuer.</1>",
-  "bugsTitle": "Bogues, commentaires et suggestions",
-  "bugsDescription": "Partagez-les sur notre serveur <0>Discord</0>.",
   "add": "Ajouter",
   "delete": "Supprimer",
   "save": "Enregistrer",
@@ -518,5 +517,8 @@
   "Authentication": "Authentification",
   "minutes": "minutes",
   "customTiles.actions": "Actions",
-  "customUpdated": "Personnalisé Mis à Jour"
+  "customUpdated": "Personnalisé Mis à Jour",
+  "updatedGameSettings": "a mis à jour les paramètres du jeu",
+  "updatedRoomSettings": "a mis à jour les paramètres de la salle",
+  "updatedMultipleSettings": "a mis à jour {{count}} paramètres"
 }
