+++ src/locales/hi/translation.json
@@ -22,6 +22,7 @@
   "tilesCount": "{{count}} टाइलें",
   "online": "ऑनलाइन",
   "setup": "गेम सेटअप",
+  "setupSubtitle": "अपना नाम दर्ज करें और सत्र शुरू करें",
   "buildGame": "गेम बनाएं",
   "access": "गेम एक्सेस",
   "setupWizard": "सेटअप विज़ार्ड",
@@ -197,15 +198,13 @@
   "pickSomething": "गेम अपडेट करने से पहले कृपया बोर्ड टैब से कम से कम एक क्रिया चुनें।",
   "appendWithAction": "जोड़ने के विकल्पों के लिए कम से कम एक स्टैंडअलोन क्रिया की आवश्यकता होती है। कृपया बोर्ड टैब से एक गैर-जोड़ने वाली क्रिया चुनें।",
   "gameDesc": "<0>ब्लिट्जेड आउट</0><1></1><2>कामुक लोगों के लिए एक ऑनलाइन वयस्क सेक्स बोर्ड गेम। एकल खेल, जोड़ों या पार्टियों के लिए डिज़ाइन किया गया।</2><3>अपना बोर्ड कस्टमाइज़ करें या किसी और का आयात करें, अपना रूम खोजें और ऑनलाइन अन्य लोगों के साथ खेलें।</3>",
+  "whySpecialTitle": "यह क्या खास बनाता है?",
+  "customizationFeature": "<0>🎲</0><1>असीमित अनुकूलन</1><2>अपनी इच्छाओं के अनुरूप कस्टम बोर्ड, क्रियाओं और तीव्रता स्तरों के साथ अपना सही अनुभव बनाएं।</2>",
+  "privacyFeature": "<0>🔒</0><1>पूर्ण गोपनीयता नियंत्रण</1><2>निजी कमरे, स्थानीय खेल विकल्प और गुमनाम सत्र। आपकी गोपनीयता और आराम पहले आता है।</2>",
+  "varietyFeature": "<0>🌟</0><1>अनंत विविधता</1><2>कोमल खोज से लेकर तीव्र रोमांच तक — समुदाय द्वारा बनाई गई सामग्री और नियमित अपडेट के साथ।</2>",
   "gettingStartedTitle": "शुरुआत करना",
   "gettingStartedMainText": "अपना सत्र शुरू करने के लिए ऊपर अपना नाम दर्ज करें। सेटअप विज़ार्ड में अपने अनुभव को अनुकूलित करें। पासा फेंकें और रोमांच शुरू होने दें।",
   "gettingStartedTip": "इसमें नए हैं? हल्के तीव्रता स्तरों से शुरुआत करें और अपनी गति से खोजें। आप हमेशा कभी भी सेटिंग्स समायोजित कर सकते हैं।",
-  "couplesTitle": "जोड़े और पार्टियां",
-  "couplesDescription": "<0>एक साथी या दोस्तों के साथ एक ही कमरे में खेल रहे हैं?</0><1>सेटिंग्स > एप्लिकेशन > निजी रूम > स्थानीय पार्टी।</1><2>नोट: यदि आप कमरे में मौजूद लोगों के साथ बातचीत नहीं करना चाहते हैं, तो सोलो चुनें।</2>",
-  "coachTitle": "रोबो कोचिंग",
-  "coachDescription": "<0>सेटिंग्स > एप्लिकेशन > मेरे रोल पढ़ें > सेव/एक्सेस गेम। फिर रोल के बगल में ऊपर का तीर क्लिक करें और अपना ऑटो रोल अंतराल चुनें।</0><1>गेम फिर अपने आप रोल करेगा और आपको करने वाली क्रिया पढ़कर सुनाएगा।</1>",
-  "bugsTitle": "बग्स, फीडबैक और सुझाव",
-  "bugsDescription": "हमारे <0>डिस्कॉर्ड सर्वर</0> पर साझा करें",
   "add": "जोड़ें",
   "delete": "हटाएं",
   "save": "सेव करें",
@@ -517,5 +516,8 @@
   "Authentication": "प्रमाणीकरण",
   "minutes": "मिनट",
   "customTiles.actions": "क्रियाएं",
-  "customUpdated": "कस्टम अपडेटेड"
+  "customUpdated": "कस्टम अपडेटेड",
+  "updatedGameSettings": "गेम सेटिंग्स अपडेट की गईं",
+  "updatedRoomSettings": "रूम सेटिंग्स अपडेट की गईं",
+  "updatedMultipleSettings": "{{count}} सेटिंग्स अपडेट की गईं"
 }
