+++ src/locales/zh/translation.json
@@ -22,6 +22,7 @@
   "tilesCount": "{{count}} 磁磚",
   "online": "線上",
   "setup": "遊戲設置",
+  "setupSubtitle": "輸入您的姓名並開始會話",
   "buildGame": "建立遊戲",
   "access": "存取遊戲",
   "setupWizard": "設置精靈",
@@ -197,15 +198,13 @@
   "pickSomething": "在更新遊戲之前，請從遊戲板標籤中至少選擇一個動作。",
   "appendWithAction": "附加選項需要至少一個獨立動作。請從遊戲板標籤中選擇一個非附加動作。",
   "gameDesc": "<0>Blitzed Out</0><1></1><2>一個為變態人士設計的線上成人性愛遊戲板遊戲。設計用於單人遊玩、情侶或聚會。</2><3>自訂您自己的遊戲板或匯入他人的遊戲板，找到您自己的房間，與其他人線上遊玩。</3>",
+  "whySpecialTitle": "有什麼特別之處？",
+  "customizationFeature": "<0>🎲</0><1>無限客製化</1><2>使用專為您的欲望量身定制的自定義遊戲板、動作和強度級別創造完美體驗。</2>",
+  "privacyFeature": "<0>🔒</0><1>完全隱私控制</1><2>私人房間、本地遊戲選項和匿名會話。您的隱私和舒適是首要考慮。</2>",
+  "varietyFeature": "<0>🌟</0><1>無限變化</1><2>從溫和探索到激烈冒險——擁有社區創建的內容和定期更新。</2>",
   "gettingStartedTitle": "開始使用",
   "gettingStartedMainText": "在上方輸入您的姓名以開始會話。在設置精靈中自定義您的體驗。擲骰子讓冒險開始。",
   "gettingStartedTip": "新手？從較輕的強度等級開始，按自己的節奏探索。您可以隨時調整設置。",
-  "couplesTitle": "情侶與聚會",
-  "couplesDescription": "<0>與伴侶或朋友在同一房間遊玩？</0><1>設定 > 應用程式 > 私人房間 > 本地聚會。</1><2>注意：如果您不想與房間中的其他人互動，請選擇單人。</2>",
-  "coachTitle": "機器人指導",
-  "coachDescription": "<0>設定 > 應用程式 > 朗讀我的擲骰 > 儲存/存取遊戲。然後點擊擲骰旁邊的上箭頭並選擇您的自動擲骰間隔。</0><1>遊戲將自動擲骰並朗讀您應該採取的動作。</1>",
-  "bugsTitle": "錯誤、回饋與建議",
-  "bugsDescription": "在我們的 <0>Discord 伺服器</0> 上分享",
   "add": "新增",
   "delete": "刪除",
   "save": "儲存",
@@ -504,5 +503,8 @@
   "Authentication": "身份驗證",
   "minutes": "分鐘",
   "customTiles.actions": "動作",
-  "customUpdated": "自定義已更新"
+  "customUpdated": "自定義已更新",
+  "updatedGameSettings": "更新了遊戲設置",
+  "updatedRoomSettings": "更新了房間設置",
+  "updatedMultipleSettings": "更新了{{count}}個設置"
 }
